requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
firebase-admin==2.9.0
google-api-core==1.0.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.8.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
pkg-resources==0.0.0
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGithub==1.37
PyJWT==1.6.0
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
